Ignore Ctrl-C when the terminal is using a key protocol
When a key protocol is in use ^C will be sent as an escape sequence, but
a raw ^C can be sent when pasting data. Pass this through, so a ^C can
be pasted and won't result in exiting insert mode.

Many terminals will strip control characters in paste data (and xterm
will strip ^C since version 388), but this provides some defense in
depth if users change settings like xterm's allowPasteControls.

Fixes #13398

Signed-off-by: David Leadbeater <dgl@dgl.cx>